src/test/java/com/leetcode/graph/hard/SwimInRisingWater_778Test.java
package com.leetcode.graph.hard;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class SwimInRisingWater_778Test {

  private final SwimInRisingWater_778 solution = new SwimInRisingWater_778();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(
          arguments(new int[][] {{0, 2}, {1, 3}}, 3),
          arguments(
              new int[][] {
                {0, 1, 2, 3, 4},
                {24, 23, 22, 21, 5},
                {12, 13, 14, 15, 16},
                {11, 17, 18, 19, 20},
                {10, 9, 8, 7, 6}
              },
              16));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void swimInWater(int[][] grid, int expectedResult) {
    int actualResult = solution.swimInWater(grid);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
